refactor(sidebar): rename ExplorerTab to ExplorerSidebar
File name now matches the exported ExplorerSidebar component; update the
protected layout import.
